+++ eventmesh-dashboard-console/src/main/java/org/apache/eventmesh/dashboard/console/EventmeshConsoleApplication.java
@@ -1,3 +1,20 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.apache.eventmesh.dashboard.console;
 
 import lombok.extern.slf4j.Slf4j;
@@ -5,18 +22,16 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
 
-/**
- * @author lambert
- */
 @Slf4j
 @SpringBootApplication
-@ComponentScan({"org.apache.eventmesh.dashboard.core"})
+@ComponentScan({"org.apache.eventmesh.dashboard.service", "org.apache.eventmesh.dashboard.console"})
 public class EventmeshConsoleApplication {
+
     public static void main(String[] args) {
-        try{
+        try {
             SpringApplication.run(EventmeshConsoleApplication.class, args);
-            log.info("{} 启动成功",EventmeshConsoleApplication.class.getSimpleName());
-        }catch (Exception e){
+            log.info("{} Successfully booted.", EventmeshConsoleApplication.class.getSimpleName());
+        } catch (Exception e) {
             log.error(e.getMessage(), e);
         }
     }
