+++ eventmesh-dashboard-core/src/main/java/org/apache/eventmesh/dashboard/core/controller/SubscriptionController.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.dashboard.console.controller;
+package org.apache.eventmesh.dashboard.core.controller;
 
 import org.apache.eventmesh.dashboard.core.dto.Result;
 import org.apache.eventmesh.dashboard.core.model.SubscriptionInfo;
